remove navigation and move out gazebo_sim
