+++ .dockerignore
@@ -1,5 +1,4 @@
 .git
-.gitmodules
 .github
 .idea
 .vscode
