+++ docker-compose.yml
@@ -1,9 +1,20 @@
 services:
+  base:
+    image: fishbot_base:latest
+    build:
+      context: .
+      dockerfile: docker/Dockerfile
+    command: ["true"]
+
   gazebo:
     container_name: fishbot_gazebo
     build:
       context: .
-      dockerfile: docker/Dockerfile
+      dockerfile: docker/Dockerfile.gazebo
+      args:
+        BASE_IMAGE: fishbot_base:latest
+    depends_on:
+      - base
     ipc: host
     tty: true
     runtime: nvidia
@@ -19,13 +30,17 @@ services:
       - /tmp/.X11-unix:/tmp/.X11-unix:rw
       - ${HOME}/.Xauthority:/root/.Xauthority:ro
     command: >
-      bash -lc "ros2 launch fishbot_description gazebo.launch.py"
+      bash -lc "ros2 launch gazebo_sim gazebo.launch.py"
 
-  navigation:
-    container_name: fishbot_navigation
+  state_publisher:
+    container_name: fishbot_state_publisher
     build:
       context: .
-      dockerfile: docker/Dockerfile
+      dockerfile: docker/Dockerfile.state
+      args:
+        BASE_IMAGE: fishbot_base:latest
+    depends_on:
+      - base
     ipc: host
     tty: true
     runtime: nvidia
@@ -41,26 +56,4 @@ services:
       - /tmp/.X11-unix:/tmp/.X11-unix:rw
       - ${HOME}/.Xauthority:/root/.Xauthority:ro
     command: >
-      bash -lc "ros2 launch fishbot_navigation2 navigation2.launch.py 'use_sim_time:=True'"
-
-  rviz:
-    container_name: fishbot_rviz
-    build:
-      context: .
-      dockerfile: docker/Dockerfile
-    ipc: host
-    tty: true
-    runtime: nvidia
-    privileged: true
-    network_mode: host
-    env_file: ./docker/ros.env
-    environment:
-      - DISPLAY=${DISPLAY:-:0}
-      - QT_X11_NO_MITSHM=1
-      - NVIDIA_VISIBLE_DEVICES=all
-      - NVIDIA_DRIVER_CAPABILITIES=all
-    volumes:
-      - /tmp/.X11-unix:/tmp/.X11-unix:rw
-      - ${HOME}/.Xauthority:/root/.Xauthority:ro
-    command: >
-      bash -lc "ros2 launch fishbot_description display_rviz2.launch.py"
+      bash -lc "ros2 launch fishbot_description state_publisher.launch.py"
