+++ docker/Dockerfile.state
@@ -1,0 +1,26 @@
+# syntax=docker/dockerfile:1
+ARG BASE_IMAGE=fishbot_base:latest
+FROM ${BASE_IMAGE}
+
+SHELL ["/bin/bash", "-c"]
+
+WORKDIR /fishbot_ws
+
+COPY src/fishbot_description/package.xml src/fishbot_description/
+
+RUN set -euxo pipefail \
+    && rosdep update \
+    && rosdep install --rosdistro humble --from-paths src --ignore-src -y
+
+COPY src/fishbot_description/ src/fishbot_description/
+
+RUN set -euxo pipefail \
+    && source /opt/ros/humble/setup.bash \
+    && colcon build --merge-install \
+        --packages-select fishbot_description
+
+COPY docker/entrypoint.sh /entrypoint.sh
+RUN chmod +x /entrypoint.sh
+
+ENTRYPOINT ["/entrypoint.sh"]
+CMD ["bash"]
