+++ src/fishbot_description/CMakeLists.txt
@@ -1,0 +1,10 @@
+cmake_minimum_required(VERSION 3.8)
+project(fishbot_description)
+
+find_package(ament_cmake REQUIRED)
+
+install(DIRECTORY launch urdf
+  DESTINATION share/${PROJECT_NAME}
+)
+
+ament_package()
