+++ src/fishbot_description/launch/state_publisher.launch.py
@@ -7,7 +7,6 @@ QQ交流群: 2642868461
 
 import os
 from launch import LaunchDescription
-from launch.substitutions import LaunchConfiguration
 from launch_ros.actions import Node
 from launch_ros.substitutions import FindPackageShare
 
@@ -33,15 +32,7 @@ def generate_launch_description():
         arguments=[urdf_model_path]
         )
 
-    rviz2_node = Node(
-        package='rviz2',
-        executable='rviz2',
-        name='rviz2',
-        output='screen',
-        )
-
-    # ld.add_action(joint_state_publisher_node)
+    ld.add_action(joint_state_publisher_node)
     ld.add_action(robot_state_publisher_node)
-    ld.add_action(rviz2_node)
 
     return ld
