+++ src/fishbot_description/package.xml
@@ -3,21 +3,16 @@
 <package format="3">
   <name>fishbot_description</name>
   <version>0.0.0</version>
-  <description>TODO: Package description</description>
+  <description>Minimal Fishbot URDF and visualization launch files.</description>
   <maintainer email="root@todo.todo">root</maintainer>
   <license>TODO: License declaration</license>
-  
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
   <depend>joint_state_publisher_gui</depend>
   <depend>joint_state_publisher</depend>
-  <depend>gazebo_ros_pkgs</depend>
-  <depend>gazebo</depend>
-  <depend>gazebo_ros</depend>
-  <depend>nav2_bringup</depend>
+  <depend>robot_state_publisher</depend>
   <export>
-    <build_type>ament_python</build_type>
+    <build_type>ament_cmake</build_type>
   </export>
 </package>
