--- src/fishbot_description/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/fishbot_description
-[install]
-install_scripts=$base/lib/fishbot_description
