--- src/fishbot_navigation2/maps/fishbot_map.yaml
@@ -1,7 +1,0 @@
-image: fishbot_map.pgm
-mode: trinary
-resolution: 0.05
-origin: [-3.37, -2.88, 0]
-negate: 0
-occupied_thresh: 0.65
-free_thresh: 0.25
