--- src/fishbot_navigation2/package.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>fishbot_navigation2</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="sangxin2014@sina.com">root</maintainer>
-  <license>TODO: License declaration</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  <exec_depend>robot_localization</exec_depend>
-  <exec_depend>nav2_bringup</exec_depend>
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
