+++ src/gazebo_sim/resource/gazebo_sim
@@ -1,0 +1,1 @@
+gazebo_sim
