+++ src/gazebo_sim/setup.cfg
@@ -1,0 +1,5 @@
+[develop]
+script_dir=$base/lib/gazebo_sim
+
+[install]
+install_scripts=$base/lib/gazebo_sim
