+++ src/gazebo_sim/setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 from glob import glob
 import os
 
-package_name = 'fishbot_description'
+package_name = 'gazebo_sim'
 
 setup(
     name=package_name,
@@ -13,19 +13,14 @@ setup(
             ['resource/' + package_name]),
         ('share/' + package_name, ['package.xml']),
         (os.path.join('share', package_name, 'launch'), glob('launch/*.launch.py')),
-        (os.path.join('share', package_name, 'urdf'), glob('urdf/**')),
-        (os.path.join('share', package_name, 'world'), glob('world/**')),
+        (os.path.join('share', package_name, 'urdf'), glob('urdf/*.urdf')),
+        (os.path.join('share', package_name, 'world'), glob('world/*.world')),
     ],
     install_requires=['setuptools'],
     zip_safe=True,
     maintainer='root',
     maintainer_email='root@todo.todo',
-    description='TODO: Package description',
+    description='Gazebo simulation assets for Fishbot',
     license='TODO: License declaration',
     tests_require=['pytest'],
-    entry_points={
-        'console_scripts': [
-            "rotate_wheel= fishbot_description.rotate_wheel:main"
-        ],
-    },
 )
