--- src/nav2_pure_pursuit_controller/nav2_pure_pursuit_controller.xml
@@ -1,10 +1,0 @@
-<class_libraries>
-    <library path="nav2_pure_pursuit_controller">
-        <class type="nav2_pure_pursuit_controller::PurePursuitController" base_class_type="nav2_core::Controller">
-            <description>
-                nav2_pure_pursuit_controller
-            </description>
-        </class>
-    </library>
-</class_libraries>
-
