--- src/nav2_pure_pursuit_controller/package.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>nav2_pure_pursuit_controller</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="87068644+fishros@users.noreply.github.com">mouse</maintainer>
-  <license>TODO: License declaration</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>nav2_common</depend>
-  <depend>nav2_core</depend>
-  <depend>nav2_util</depend>
-  <depend>nav2_costmap_2d</depend>
-  <depend>rclcpp</depend>
-  <depend>geometry_msgs</depend>
-  <depend>nav2_msgs</depend>
-  <depend>pluginlib</depend>
-  <depend>tf2</depend>
-  <depend>tf2_geometry_msgs</depend>
-
-  <test_depend>ament_cmake_gtest</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-    <nav2_core plugin="${prefix}/nav2_pure_pursuit_controller.xml" />
-  </export>
-
-</package>
